// tests/pages/enums/ProductNames.ts
export enum ProductNames {
    AdvancedHipJointSupportSmallMedium = "Advanced Hip & Joint Support for Small & Medium Breeds",
    AdvancedHipJointSupportforLargeBreeds = "Advanced Hip & Joint Support for Large Breeds",
    WildAlaskanPollockOil8oz = "Wild Alaskan Pollock Oil 8oz",
    WildAlaskanPollockOil16oz = "Wild Alaskan Pollock Oil 16oz"
}